Move display strings into new lang file
