+++ src/components/formatting/ResponseAdherence.tsx
@@ -1,15 +1,9 @@
 import React from 'react';
 import { Table } from 'reactstrap';
 import * as colors from '../../colors';
+import { CONDITIONAL_FORMATTING_TITLE, FI_RESPONSE_ADHERENCE_TITLE } from '../../configs/lang';
 import { GREEN_THRESHOLD, ORANGE_THRESHOLD, YELLOW_THRESHOLD } from '../../configs/settings';
-import {
-  CONDITIONAL_FORMATTING_TITLE,
-  FI_RESPONSE_ADHERENCE_TITLE,
-  GREEN,
-  ORANGE,
-  RED,
-  YELLOW,
-} from '../../constants';
+import { GREEN, ORANGE, RED, YELLOW } from '../../constants';
 import { percentage } from '../../helpers/utils';
 
 /** Displays conditional formatting guide for Focus Investigation Response Adherence */
