+++ src/components/page/Header/index.tsx
@@ -23,35 +23,36 @@ import {
   ENABLE_PRACTITIONERS,
   ENABLE_TEAMS,
   ENABLE_USERS,
-  OPENSRP_LOGOUT_URL,
   WEBSITE_NAME,
 } from '../../../configs/env';
 import {
   ABOUT,
   ADMIN,
   ASSIGN,
-  ASSIGN_PLAN_URL,
-  FI_URL,
   FOCUS_INVESTIGATION,
   HOME,
-  INTERVENTION_IRS_DRAFTS_URL,
-  INTERVENTION_IRS_URL,
   IRS_REPORTING_TITLE,
   IRS_TITLE,
   LOGIN,
-  LOGIN_URL,
-  LOGOUT_URL,
   MONITOR,
   ORGANIZATIONS_LABEL,
-  ORGANIZATIONS_LIST_URL,
-  PLAN_LIST_URL,
   PLAN_TITLE,
   PLANS,
   PRACTITIONERS,
-  PRACTITIONERS_LIST_URL,
-  REPORT_IRS_PLAN_URL,
   SIGN_OUT,
   USERS,
+} from '../../../configs/lang';
+import {
+  ASSIGN_PLAN_URL,
+  FI_URL,
+  INTERVENTION_IRS_DRAFTS_URL,
+  INTERVENTION_IRS_URL,
+  LOGIN_URL,
+  LOGOUT_URL,
+  ORGANIZATIONS_LIST_URL,
+  PLAN_LIST_URL,
+  PRACTITIONERS_LIST_URL,
+  REPORT_IRS_PLAN_URL,
 } from '../../../constants';
 import './Header.css';
