+++ src/configs/settings.ts
@@ -16,29 +16,45 @@ import { Providers } from '@onaio/gatekeeper';
 import { Color } from 'csstype';
 import { Expression, LngLatBoundsLike } from 'mapbox-gl';
 import {
-  A1,
-  A2,
-  B1,
-  B2,
   BCC_ACTIVITY,
   BCC_ACTIVITY_DESCRIPTION,
-  BCC_CODE,
   BCC_GOAL_DESCRIPTION,
   BCC_GOAL_MEASURE,
   BEDNET_ACTIVITY,
   BEDNET_ACTIVITY_DESCRIPTION,
-  BEDNET_DISTRIBUTION_ACTIVITY_CODE,
-  BEDNET_DISTRIBUTION_CODE,
   BEDNET_GOAL_MEASURE,
   BLOOD_SCREENING_ACTIVITY,
-  BLOOD_SCREENING_ACTIVITY_CODE,
   BLOOD_SCREENING_ACTIVITY_DESCRIPTION,
-  BLOOD_SCREENING_CODE,
   BLOOD_SCREENING_GOAL_MEASURE,
   CASE_CONFIRMATION_ACTIVITY,
-  CASE_CONFIRMATION_ACTIVITY_CODE,
   CASE_CONFIRMATION_ACTIVITY_DESCRIPTION,
   CASE_CONFIRMATION_ACTIVITY_GOAL_MEASURE,
+  INVESTIGATION,
+  IRS_ACTIVITY,
+  IRS_ACTIVITY_DESCRIPTION,
+  IRS_GOAL_DESCRIPTION,
+  IRS_GOAL_MEASURE,
+  LARVAL_DIPPING_ACTIVITY,
+  LARVAL_DIPPING_ACTIVITY_DESCRIPTION,
+  LARVAL_DIPPING_GOAL_MEASURE,
+  MOSQUITO_COLLECTION_ACTIVITY,
+  MOSQUITO_COLLECTION_ACTIVITY_DESCRIPTION,
+  MOSQUITO_COLLECTION_GOAL_MEASURE,
+  RACD_ACTIVITY,
+  RACD_ACTIVITY_DESCRIPTION,
+  RACD_ACTIVITY_GOAL_MEASURE,
+} from '../configs/lang';
+import {
+  A1,
+  A2,
+  B1,
+  B2,
+  BCC_CODE,
+  BEDNET_DISTRIBUTION_ACTIVITY_CODE,
+  BEDNET_DISTRIBUTION_CODE,
+  BLOOD_SCREENING_ACTIVITY_CODE,
+  BLOOD_SCREENING_CODE,
+  CASE_CONFIRMATION_ACTIVITY_CODE,
   CASE_CONFIRMATION_CODE,
   CASE_NUMBER_CODE,
   CASE_TRIGGERED,
@@ -48,28 +64,14 @@ import {
   FI_STATUS_CODE,
   HIGH_PRIORITIY,
   INTERVENTION_TYPE_CODE,
-  INVESTIGATION,
-  IRS_ACTIVITY,
-  IRS_ACTIVITY_DESCRIPTION,
   IRS_CODE,
-  IRS_GOAL_DESCRIPTION,
-  IRS_GOAL_MEASURE,
-  LARVAL_DIPPING_ACTIVITY,
   LARVAL_DIPPING_ACTIVITY_CODE,
-  LARVAL_DIPPING_ACTIVITY_DESCRIPTION,
   LARVAL_DIPPING_CODE,
-  LARVAL_DIPPING_GOAL_MEASURE,
   LOW_PRIORITY,
   MEDIUM_PRIORITY,
-  MOSQUITO_COLLECTION_ACTIVITY,
   MOSQUITO_COLLECTION_ACTIVITY_CODE,
-  MOSQUITO_COLLECTION_ACTIVITY_DESCRIPTION,
   MOSQUITO_COLLECTION_CODE,
-  MOSQUITO_COLLECTION_GOAL_MEASURE,
   OPENSRP_EVENT_ID_CODE,
-  RACD_ACTIVITY,
-  RACD_ACTIVITY_DESCRIPTION,
-  RACD_ACTIVITY_GOAL_MEASURE,
   RACD_REGISTER_FAMILY_CODE,
   ROUTINE,
   TASK_GENERATION_STATUS_CODE,
