+++ src/containers/pages/IRS/assignments/index.tsx
@@ -9,18 +9,16 @@ import { Store } from 'redux';
 import HeaderBreadcrumb from '../../../../components/page/HeaderBreadcrumb/HeaderBreadcrumb';
 import Loading from '../../../../components/page/Loading';
 import {
-  ASSIGN_PLAN_URL,
   ASSIGN_PLANS,
   END_DATE,
   HOME,
-  HOME_URL,
   INTERVENTION,
   NO_PLANS_LOADED_MESSAGE,
   PLAN_STATUS,
-  REPORT_IRS_PLAN_URL,
   START_DATE,
   TITLE,
-} from '../../../../constants';
+} from '../../../../configs/lang';
+import { ASSIGN_PLAN_URL, HOME_URL, REPORT_IRS_PLAN_URL } from '../../../../constants';
 import { OpenSRPService } from '../../../../services/opensrp';
 import IRSPlansReducer, {
   reducerName as IRSPlansReducerName,
