+++ src/containers/pages/InterventionPlan/IRS/plan/index.tsx
@@ -22,28 +22,13 @@ import {
 import {
   ADMIN_LEVEL,
   ASSIGN_JURISDICTIONS,
-  ASSIGN_PLAN_URL,
   ASSIGN_PLANS,
-  DRAFT,
   DRAFT_PLAN_TITLE,
   HOME,
-  HOME_URL,
-  INTERVENTION_IRS_URL,
   IRS_TITLE,
   JURISDICTION,
-  JURISDICTION_ID,
-  MAP_ID,
   NAME,
-  NEW,
   NEW_PLAN,
-  OPENSRP_FIND_BY_PROPERTIES,
-  OPENSRP_GET_ASSIGNMENTS_ENDPOINT,
-  OPENSRP_LOCATION,
-  OPENSRP_ORGANIZATION_ENDPOINT,
-  OPENSRP_PARENT_ID,
-  OPENSRP_PLANS,
-  OPENSRP_POST_ASSIGNMENTS_ENDPOINT,
-  PARENTID,
   SAVE_AS_DRAFT,
   SAVE_ASSIGNMENTS,
   SAVE_FINALIZED_PLAN,
@@ -53,6 +38,23 @@ import {
   SPRAY_AREA_HEADER,
   TEAMS_ASSIGNMENT,
   TYPE_LABEL,
+} from '../../../../../configs/lang';
+import {
+  ASSIGN_PLAN_URL,
+  DRAFT,
+  HOME_URL,
+  INTERVENTION_IRS_URL,
+  JURISDICTION_ID,
+  MAP_ID,
+  NEW,
+  OPENSRP_FIND_BY_PROPERTIES,
+  OPENSRP_GET_ASSIGNMENTS_ENDPOINT,
+  OPENSRP_LOCATION,
+  OPENSRP_ORGANIZATION_ENDPOINT,
+  OPENSRP_PARENT_ID,
+  OPENSRP_PLANS,
+  OPENSRP_POST_ASSIGNMENTS_ENDPOINT,
+  PARENTID,
 } from '../../../../../constants';
 import {
   FlexObject,
