+++ src/containers/pages/OrganizationViews/AssignPractitioners/index.tsx
@@ -24,20 +24,22 @@ import {
   ASSIGN,
   ASSIGN_PRACTITIONERS,
   ASSIGN_PRACTITIONERS_TO_ORG,
-  ASSIGN_PRACTITIONERS_URL,
   DISCARD_CHANGES,
   HOME,
-  HOME_URL,
   NO_PRACTITIONERS_ADDED_YET,
   ON_REROUTE_WITH_UNSAVED_CHANGES,
-  OPENSRP_ADD_PRACTITIONER_ROLE_ENDPOINT,
-  OPENSRP_PRACTITIONER_ENDPOINT,
   ORGANIZATIONS_LABEL,
-  ORGANIZATIONS_LIST_URL,
-  PRACTITIONER_CODE,
   PRACTITIONERS,
   PRACTITIONERS_ASSIGNED_TO_ORG,
   SAVE_CHANGES,
+} from '../../../../configs/lang';
+import {
+  ASSIGN_PRACTITIONERS_URL,
+  HOME_URL,
+  OPENSRP_ADD_PRACTITIONER_ROLE_ENDPOINT,
+  OPENSRP_PRACTITIONER_ENDPOINT,
+  ORGANIZATIONS_LIST_URL,
+  PRACTITIONER_CODE,
   SINGLE_ORGANIZATION_URL,
 } from '../../../../constants';
 import { useConfirmOnBrowserUnload } from '../../../../helpers/hooks';
